+++ Tests.Integration.Presentation.Web/Tools/OrganizationUnitHelper.cs
@@ -47,9 +47,10 @@ namespace Tests.Integration.Presentation.Web.Tools
         public static async Task<List<UnitWithAccessRightsDTO>> GetUnitAccessRightsForOrganization(Guid organizationUuid, Cookie optionalLogin = null)
         {
             var cookie = optionalLogin ?? await HttpApi.GetCookieAsync(OrganizationRole.GlobalAdmin);
-            var orgUnitUrl = TestEnvironment.CreateUrl($"api/v1/organizations/{organizationUuid}/organization-units/access-rights");
+            var orgUnitUrl = TestEnvironment.CreateUrl($"api/v1/organizations/{organizationUuid}/organization-units/all/access-rights");
 
             using var response = await HttpApi.GetWithCookieAsync(orgUnitUrl, cookie);
+            var res = await response.Content.ReadAsStringAsync();
             Assert.Equal(HttpStatusCode.OK, response.StatusCode);
 
             return await response.ReadResponseBodyAsKitosApiResponseAsync<List<UnitWithAccessRightsDTO>>();
